requirements.txt
aiofiles==23.2.1
aiosqlite==0.20.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.1.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.103.2
greenlet==3.0.3
h11==0.14.0
idna==3.6
orjson==3.9.14
passlib==1.7.4
pydantic==1.10.14
PyJWT==2.6.0
python-dotenv==1.0.1
python-multipart==0.0.9
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.23.2
